Additional size_t corrections
I didn't catch these on first pass.

+++ src/test/Helpers.h
@@ -366,7 +366,7 @@ class PixelEqHelper {
     Message msg;
     bool failed = false;
 
-    for ( int i = 0; i < CompoundNumChannels<PixelT>::value; i++ ) {
+    for ( size_t i = 0; i < CompoundNumChannels<PixelT>::value; i++ ) {
       const std::string idx = "["+stringify(i)+"]";
       AssertionResult ret = comp_helper( ename+idx, aname+idx,
                                          expected.impl()(i), actual.impl()(i) );
